Using strings, each method can show its progress as it sorts
We used strings that were added to after each swap, and then at the end of the swap, we added another string that stores the array for that cycle, and it repeats until it is done. For QuickSort, there is only one cycle, so I stored each of the swaps and added more details about them in each string, and then for the final string i added the entire finsihed array.